% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whed_wide.R
\name{whed_wide}
\alias{whed_wide}
\title{Transform WHED Data to Wide Format}
\usage{
whed_wide()
}
\value{
A data frame in wide format with individual columns for each degree and division.
}
\description{
This function transforms the WHED dataset into a wide format with individual columns for each degree and division.
}
\examples{
# Transform the WHED dataset back to wide format
whed_wide <- whed_wide()

}
