% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whed.R
\docType{data}
\name{whed}
\alias{whed}
\title{WHED Dataset}
\format{
A data frame with 21165 rows and 271 variables:
\describe{
\item{name}{Name of the institution}
\item{url}{Website URL of the institution}
\item{country}{Country where the institution is located}
\item{id}{Unique identifier for the institution}
\item{city}{City where the institution is located}
\item{name_detail}{Detailed name of the institution}
\item{funding_type}{Type of funding for the institution (e.g., public, private)}
\item{languages}{Languages used at the institution}
\item{accrediting_agency}{Accrediting agency for the institution}
\item{year_founded}{Year the institution was founded}
\item{academic_year}{Academic year details}
\item{business_school}{Indicates if the institution has a business school}
\item{degree_count}{Number of degrees offered}
\item{division_name_X}{Name of division X}
\item{division_fields_study_X}{Fields of study in division X}
\item{degree_name_X}{Name of degree X}
\item{degree_fields_study_X}{Fields of study for degree X}
}
}
\source{
Source of the data: \url{https://whed.net/home.php}
}
\usage{
whed
}
\description{
This dataset contains institution-level information on higher education institutions around the world.
}
\keyword{datasets}
