% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whed.R
\docType{data}
\name{whed}
\alias{whed}
\title{WHED Dataset}
\format{
A data frame with X rows and Y variables:
\describe{
\item{column1}{Description of column1}
\item{column2}{Description of column2}
...
}
}
\source{
Source of the data
}
\usage{
whed
}
\description{
This dataset contains information from the WHED dataset.
}
\keyword{datasets}
